fix(web): harden and modularize Zendesk connectors
Prevent unsafe Zendesk tool argument coercion and third-party requester impersonation while keeping connector test behavior easier to extend. Document the stable gateway token design and split the connector test and Zendesk logic into smaller modules for maintenance.
