+++ apps/web/src/lib/connectors/config-validation.ts
@@ -1,0 +1,3 @@
+export type ConnectorConfigValidationResult =
+  | { valid: true }
+  | { valid: false; missing?: string[]; message?: string }
